Anything in the file runs now and it is aweesome

+++ sql/schema.sql
@@ -1,5 +1,5 @@
-CREATE TABLE IF NOT EXISTS classroom(
+CREATE TABLE IF NOT EXISTS classroom (
     building       varchar(15),
     room_number    varchar(7),
-    capacity       numeric(4,0),
+    capacity       int
 );
